requirements.txt
pandas==1.4.0
pytest==7.1.*
autopep8==2.0.1
entrezpy==2.1.3 # NCBI API
biopython==1.80